requirements.txt
langchain==0.3.26
langchain_core==0.3.67
langchain_google_genai==2.1.6
langchain_neo4j==0.4.0
langchain_openai==0.3.27
langgraph==0.5.0
neo4j==5.28.1
python-dotenv==1.1.1
typing_extensions==4.14.0
